Flexible key manager policy signers
