+++ .changelog/2444.feature.md
@@ -1,0 +1,5 @@
+Flexible key manager policy signers
+
+The key manager runtime has been split into multiple crates to make its code
+reusable.  It is now possible for others to write their own key managers that
+use a different set of trusted policy signers.
