+++ keymanager-api-common/Cargo.toml
@@ -1,12 +1,11 @@
 [package]
-name = "oasis-core-keymanager-api"
+name = "oasis-core-keymanager-api-common"
 version = "0.3.0-alpha"
 authors = ["Oasis Labs Inc. <info@oasislabs.com>"]
 
 [dependencies]
 base64 = "0.10.1"
-oasis-core-keymanager-api-common = { path = "../../keymanager-api-common" }
-oasis-core-runtime = { path = "../../runtime" }
+oasis-core-runtime = { path = "../runtime" }
 serde = "1.0.71"
 serde_derive = "1.0"
 serde_bytes = "~0.10"
