+++ keymanager-lib/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "oasis-core-keymanager-lib"
+version = "0.3.0-alpha"
+authors = ["Oasis Labs Inc. <info@oasislabs.com>"]
+edition = "2018"
+
+[dependencies]
+oasis-core-runtime = { path = "../runtime" }
+oasis-core-keymanager-api-common = { path = "../keymanager-api-common" }
+oasis-core-keymanager-client = { path = "../keymanager-client" }
+failure = "0.1.5"
+lazy_static = "1.3.0"
+lru = "0.1.17"
+io-context = "0.2.0"
+rand = "0.6.5"
+sgx-isa = { version = "0.3.0", features = ["sgxstd"] }
+sp800-185 = "0.2.0"
+tiny-keccak = "1.4.2"
+x25519-dalek = "0.5.1"
+zeroize = "0.6"
