+++ keymanager-lib/src/context.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 
 use oasis_core_runtime::{common::runtime::RuntimeId, Protocol};
 
-pub(crate) struct Context {
+pub struct Context {
     pub runtime_id: RuntimeId,
     pub protocol: Arc<Protocol>,
 }
