+++ keymanager-lib/src/lib.rs
@@ -1,0 +1,16 @@
+extern crate failure;
+extern crate io_context;
+extern crate lazy_static;
+extern crate lru;
+extern crate oasis_core_keymanager_api_common;
+extern crate oasis_core_keymanager_client;
+extern crate oasis_core_runtime;
+extern crate rand;
+extern crate sp800_185;
+extern crate tiny_keccak;
+extern crate x25519_dalek;
+extern crate zeroize;
+
+pub mod context;
+pub mod kdf;
+pub mod policy;
