+++ keymanager-runtime/src/methods.rs
@@ -3,7 +3,7 @@ use failure::Fallible;
 use oasis_core_keymanager_api::*;
 use oasis_core_runtime::rpc::Context as RpcContext;
 
-use crate::{kdf::Kdf, policy::Policy};
+use oasis_core_keymanager_lib::{kdf::Kdf, policy::Policy};
 
 /// See `Kdf::get_or_create_keys`.
 pub fn get_or_create_keys(req: &RequestIds, ctx: &mut RpcContext) -> Fallible<ContractKey> {
